moving transaction tests to e2e
