+++ lib/utils.go
@@ -11,11 +11,9 @@ package main
 
 import "C"
 import (
-	"context"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"math/big"
 	"os"
 	"path/filepath"
 	"reflect"
@@ -24,8 +22,6 @@ import (
 	"testing"
 	"time"
 
-	gethcommon "github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core"
 	gethparams "github.com/ethereum/go-ethereum/params"
 	"github.com/stretchr/testify/require"
@@ -512,502 +508,6 @@ func testAccountLogout(t *testing.T) bool {
 	// now, logout and check if identity is removed indeed
 	if whisperService.HasKeyPair(pubKey) {
 		t.Error("identity not cleared from whisper")
-		return false
-	}
-
-	return true
-}
-
-func testCompleteTransaction(t *testing.T) bool {
-	txQueueManager := statusAPI.TxQueueManager()
-	txQueue := txQueueManager.TransactionQueue()
-
-	txQueue.Reset()
-	EnsureNodeSync(statusAPI.NodeManager())
-
-	// log into account from which transactions will be sent
-	if err := statusAPI.SelectAccount(TestConfig.Account1.Address, TestConfig.Account1.Password); err != nil {
-		t.Errorf("cannot select account: %v. Error %q", TestConfig.Account1.Address, err)
-		return false
-	}
-
-	// make sure you panic if transaction complete doesn't return
-	queuedTxCompleted := make(chan struct{}, 1)
-	abortPanic := make(chan struct{}, 1)
-	common.PanicAfter(10*time.Second, abortPanic, "testCompleteTransaction")
-
-	// replace transaction notification handler
-	var txHash = ""
-	signal.SetDefaultNodeNotificationHandler(func(jsonEvent string) {
-		var envelope signal.Envelope
-		if err := json.Unmarshal([]byte(jsonEvent), &envelope); err != nil {
-			t.Errorf("cannot unmarshal event's JSON: %s. Error %q", jsonEvent, err)
-			return
-		}
-		if envelope.Type == txqueue.EventTransactionQueued {
-			event := envelope.Event.(map[string]interface{})
-			t.Logf("transaction queued (will be completed shortly): {id: %s}\n", event["id"].(string))
-
-			completeTxResponse := common.CompleteTransactionResult{}
-			rawResponse := CompleteTransaction(C.CString(event["id"].(string)), C.CString(TestConfig.Account1.Password))
-
-			if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &completeTxResponse); err != nil {
-				t.Errorf("cannot decode RecoverAccount response (%s): %v", C.GoString(rawResponse), err)
-			}
-
-			if completeTxResponse.Error != "" {
-				t.Errorf("cannot complete queued transaction[%v]: %v", event["id"], completeTxResponse.Error)
-			}
-
-			txHash = completeTxResponse.Hash
-
-			t.Logf("transaction complete: https://testnet.etherscan.io/tx/%s", txHash)
-			abortPanic <- struct{}{} // so that timeout is aborted
-			queuedTxCompleted <- struct{}{}
-		}
-	})
-
-	// this call blocks, up until Complete Transaction is called
-	txCheckHash, err := statusAPI.SendTransaction(context.TODO(), common.SendTxArgs{
-		From:  common.FromAddress(TestConfig.Account1.Address),
-		To:    common.ToAddress(TestConfig.Account2.Address),
-		Value: (*hexutil.Big)(big.NewInt(1000000000000)),
-	})
-	if err != nil {
-		t.Errorf("Failed to SendTransaction: %s", err)
-		return false
-	}
-
-	<-queuedTxCompleted // make sure that complete transaction handler completes its magic, before we proceed
-
-	if txHash != txCheckHash.Hex() {
-		t.Errorf("Transaction hash returned from SendTransaction is invalid: expected %s, got %s",
-			txCheckHash.Hex(), txHash)
-		return false
-	}
-
-	if reflect.DeepEqual(txCheckHash, gethcommon.Hash{}) {
-		t.Error("Test failed: transaction was never queued or completed")
-		return false
-	}
-
-	if txQueue.Count() != 0 {
-		t.Error("tx queue must be empty at this point")
-		return false
-	}
-
-	return true
-}
-
-func testCompleteMultipleQueuedTransactions(t *testing.T) bool { //nolint: gocyclo
-	txQueue := statusAPI.TxQueueManager().TransactionQueue()
-	txQueue.Reset()
-
-	// log into account from which transactions will be sent
-	if err := statusAPI.SelectAccount(TestConfig.Account1.Address, TestConfig.Account1.Password); err != nil {
-		t.Errorf("cannot select account: %v", TestConfig.Account1.Address)
-		return false
-	}
-
-	// make sure you panic if transaction complete doesn't return
-	testTxCount := 3
-	txIDs := make(chan string, testTxCount)
-	allTestTxCompleted := make(chan struct{}, 1)
-
-	// replace transaction notification handler
-	signal.SetDefaultNodeNotificationHandler(func(jsonEvent string) {
-		var txID string
-		var envelope signal.Envelope
-		if err := json.Unmarshal([]byte(jsonEvent), &envelope); err != nil {
-			t.Errorf("cannot unmarshal event's JSON: %s", jsonEvent)
-			return
-		}
-		if envelope.Type == txqueue.EventTransactionQueued {
-			event := envelope.Event.(map[string]interface{})
-			txID = event["id"].(string)
-			t.Logf("transaction queued (will be completed in a single call, once aggregated): {id: %s}\n", txID)
-
-			txIDs <- txID
-		}
-	})
-
-	//  this call blocks, and should return when DiscardQueuedTransaction() for a given tx id is called
-	sendTx := func() {
-		txHashCheck, err := statusAPI.SendTransaction(context.TODO(), common.SendTxArgs{
-			From:  common.FromAddress(TestConfig.Account1.Address),
-			To:    common.ToAddress(TestConfig.Account2.Address),
-			Value: (*hexutil.Big)(big.NewInt(1000000000000)),
-		})
-		if err != nil {
-			t.Errorf("unexpected error thrown: %v", err)
-			return
-		}
-
-		if reflect.DeepEqual(txHashCheck, gethcommon.Hash{}) {
-			t.Error("transaction returned empty hash")
-			return
-		}
-	}
-
-	// wait for transactions, and complete them in a single call
-	completeTxs := func(txIDStrings string) {
-		var parsedIDs []string
-		if err := json.Unmarshal([]byte(txIDStrings), &parsedIDs); err != nil {
-			t.Error(err)
-			return
-		}
-
-		parsedIDs = append(parsedIDs, "invalid-tx-id")
-		updatedTxIDStrings, _ := json.Marshal(parsedIDs)
-
-		// complete
-		resultsString := CompleteTransactions(C.CString(string(updatedTxIDStrings)), C.CString(TestConfig.Account1.Password))
-		resultsStruct := common.CompleteTransactionsResult{}
-		if err := json.Unmarshal([]byte(C.GoString(resultsString)), &resultsStruct); err != nil {
-			t.Error(err)
-			return
-		}
-		results := resultsStruct.Results
-
-		if len(results) != (testTxCount+1) || results["invalid-tx-id"].Error != txqueue.ErrQueuedTxIDNotFound.Error() {
-			t.Errorf("cannot complete txs: %v", results)
-			return
-		}
-		for txID, txResult := range results {
-			if txID != txResult.ID {
-				t.Errorf("tx id not set in result: expected id is %s", txID)
-				return
-			}
-			if txResult.Error != "" && txID != "invalid-tx-id" {
-				t.Errorf("invalid error for %s", txID)
-				return
-			}
-			if txResult.Hash == zeroHash && txID != "invalid-tx-id" {
-				t.Errorf("invalid hash (expected non empty hash): %s", txID)
-				return
-			}
-
-			if txResult.Hash != zeroHash {
-				t.Logf("transaction complete: https://testnet.etherscan.io/tx/%s", txResult.Hash)
-			}
-		}
-
-		time.Sleep(1 * time.Second) // make sure that tx complete signal propagates
-		for _, txID := range parsedIDs {
-			if txQueue.Has(common.QueuedTxID(txID)) {
-				t.Errorf("txqueue should not have test tx at this point (it should be completed): %s", txID)
-				return
-			}
-		}
-	}
-	go func() {
-		var txIDStrings []string
-		for i := 0; i < testTxCount; i++ {
-			txIDStrings = append(txIDStrings, <-txIDs)
-		}
-
-		txIDJSON, _ := json.Marshal(txIDStrings)
-		completeTxs(string(txIDJSON))
-		allTestTxCompleted <- struct{}{}
-	}()
-
-	// send multiple transactions
-	for i := 0; i < testTxCount; i++ {
-		go sendTx()
-	}
-
-	select {
-	case <-allTestTxCompleted:
-		// pass
-	case <-time.After(20 * time.Second):
-		t.Error("test timed out")
-		return false
-	}
-
-	if txQueue.Count() != 0 {
-		t.Error("tx queue must be empty at this point")
-		return false
-	}
-
-	return true
-}
-
-func testDiscardTransaction(t *testing.T) bool { //nolint: gocyclo
-	txQueue := statusAPI.TxQueueManager().TransactionQueue()
-	txQueue.Reset()
-
-	// log into account from which transactions will be sent
-	if err := statusAPI.SelectAccount(TestConfig.Account1.Address, TestConfig.Account1.Password); err != nil {
-		t.Errorf("cannot select account: %v", TestConfig.Account1.Address)
-		return false
-	}
-
-	// make sure you panic if transaction complete doesn't return
-	completeQueuedTransaction := make(chan struct{}, 1)
-	common.PanicAfter(20*time.Second, completeQueuedTransaction, "testDiscardTransaction")
-
-	// replace transaction notification handler
-	var txID string
-	txFailedEventCalled := false
-	signal.SetDefaultNodeNotificationHandler(func(jsonEvent string) {
-		var envelope signal.Envelope
-		if err := json.Unmarshal([]byte(jsonEvent), &envelope); err != nil {
-			t.Errorf("cannot unmarshal event's JSON: %s", jsonEvent)
-			return
-		}
-		if envelope.Type == txqueue.EventTransactionQueued {
-			event := envelope.Event.(map[string]interface{})
-			txID = event["id"].(string)
-			t.Logf("transaction queued (will be discarded soon): {id: %s}\n", txID)
-
-			if !txQueue.Has(common.QueuedTxID(txID)) {
-				t.Errorf("txqueue should still have test tx: %s", txID)
-				return
-			}
-
-			// discard
-			discardResponse := common.DiscardTransactionResult{}
-			rawResponse := DiscardTransaction(C.CString(txID))
-
-			if err := json.Unmarshal([]byte(C.GoString(rawResponse)), &discardResponse); err != nil {
-				t.Errorf("cannot decode RecoverAccount response (%s): %v", C.GoString(rawResponse), err)
-			}
-
-			if discardResponse.Error != "" {
-				t.Errorf("cannot discard tx: %v", discardResponse.Error)
-				return
-			}
-
-			// try completing discarded transaction
-			_, err := statusAPI.CompleteTransaction(common.QueuedTxID(txID), TestConfig.Account1.Password)
-			if err != txqueue.ErrQueuedTxIDNotFound {
-				t.Error("expects tx not found, but call to CompleteTransaction succeeded")
-				return
-			}
-
-			time.Sleep(1 * time.Second) // make sure that tx complete signal propagates
-			if txQueue.Has(common.QueuedTxID(txID)) {
-				t.Errorf("txqueue should not have test tx at this point (it should be discarded): %s", txID)
-				return
-			}
-
-			completeQueuedTransaction <- struct{}{} // so that timeout is aborted
-		}
-
-		if envelope.Type == txqueue.EventTransactionFailed {
-			event := envelope.Event.(map[string]interface{})
-			t.Logf("transaction return event received: {id: %s}\n", event["id"].(string))
-
-			receivedErrMessage := event["error_message"].(string)
-			expectedErrMessage := txqueue.ErrQueuedTxDiscarded.Error()
-			if receivedErrMessage != expectedErrMessage {
-				t.Errorf("unexpected error message received: got %v", receivedErrMessage)
-				return
-			}
-
-			receivedErrCode := event["error_code"].(string)
-			if receivedErrCode != txqueue.SendTransactionDiscardedErrorCode {
-				t.Errorf("unexpected error code received: got %v", receivedErrCode)
-				return
-			}
-
-			txFailedEventCalled = true
-		}
-	})
-
-	// this call blocks, and should return when DiscardQueuedTransaction() is called
-	txHashCheck, err := statusAPI.SendTransaction(context.TODO(), common.SendTxArgs{
-		From:  common.FromAddress(TestConfig.Account1.Address),
-		To:    common.ToAddress(TestConfig.Account2.Address),
-		Value: (*hexutil.Big)(big.NewInt(1000000000000)),
-	})
-	if err != txqueue.ErrQueuedTxDiscarded {
-		t.Errorf("expected error not thrown: %v", err)
-		return false
-	}
-
-	if !reflect.DeepEqual(txHashCheck, gethcommon.Hash{}) {
-		t.Error("transaction returned hash, while it shouldn't")
-		return false
-	}
-
-	if txQueue.Count() != 0 {
-		t.Error("tx queue must be empty at this point")
-		return false
-	}
-
-	if !txFailedEventCalled {
-		t.Error("expected tx failure signal is not received")
-		return false
-	}
-
-	return true
-}
-
-func testDiscardMultipleQueuedTransactions(t *testing.T) bool { //nolint: gocyclo
-	txQueue := statusAPI.TxQueueManager().TransactionQueue()
-	txQueue.Reset()
-
-	// log into account from which transactions will be sent
-	if err := statusAPI.SelectAccount(TestConfig.Account1.Address, TestConfig.Account1.Password); err != nil {
-		t.Errorf("cannot select account: %v", TestConfig.Account1.Address)
-		return false
-	}
-
-	// make sure you panic if transaction complete doesn't return
-	testTxCount := 3
-	txIDs := make(chan string, testTxCount)
-	allTestTxDiscarded := make(chan struct{}, 1)
-
-	// replace transaction notification handler
-	txFailedEventCallCount := 0
-	signal.SetDefaultNodeNotificationHandler(func(jsonEvent string) {
-		var txID string
-		var envelope signal.Envelope
-		if err := json.Unmarshal([]byte(jsonEvent), &envelope); err != nil {
-			t.Errorf("cannot unmarshal event's JSON: %s", jsonEvent)
-			return
-		}
-		if envelope.Type == txqueue.EventTransactionQueued {
-			event := envelope.Event.(map[string]interface{})
-			txID = event["id"].(string)
-			t.Logf("transaction queued (will be discarded soon): {id: %s}\n", txID)
-
-			if !txQueue.Has(common.QueuedTxID(txID)) {
-				t.Errorf("txqueue should still have test tx: %s", txID)
-				return
-			}
-
-			txIDs <- txID
-		}
-
-		if envelope.Type == txqueue.EventTransactionFailed {
-			event := envelope.Event.(map[string]interface{})
-			t.Logf("transaction return event received: {id: %s}\n", event["id"].(string))
-
-			receivedErrMessage := event["error_message"].(string)
-			expectedErrMessage := txqueue.ErrQueuedTxDiscarded.Error()
-			if receivedErrMessage != expectedErrMessage {
-				t.Errorf("unexpected error message received: got %v", receivedErrMessage)
-				return
-			}
-
-			receivedErrCode := event["error_code"].(string)
-			if receivedErrCode != txqueue.SendTransactionDiscardedErrorCode {
-				t.Errorf("unexpected error code received: got %v", receivedErrCode)
-				return
-			}
-
-			txFailedEventCallCount++
-			if txFailedEventCallCount == testTxCount {
-				allTestTxDiscarded <- struct{}{}
-			}
-		}
-	})
-
-	// this call blocks, and should return when DiscardQueuedTransaction() for a given tx id is called
-	sendTx := func() {
-		txHashCheck, err := statusAPI.SendTransaction(context.TODO(), common.SendTxArgs{
-			From:  common.FromAddress(TestConfig.Account1.Address),
-			To:    common.ToAddress(TestConfig.Account2.Address),
-			Value: (*hexutil.Big)(big.NewInt(1000000000000)),
-		})
-		if err != txqueue.ErrQueuedTxDiscarded {
-			t.Errorf("expected error not thrown: %v", err)
-			return
-		}
-
-		if !reflect.DeepEqual(txHashCheck, gethcommon.Hash{}) {
-			t.Error("transaction returned hash, while it shouldn't")
-			return
-		}
-	}
-
-	// wait for transactions, and discard immediately
-	discardTxs := func(txIDStrings string) {
-		var parsedIDs []string
-		if err := json.Unmarshal([]byte(txIDStrings), &parsedIDs); err != nil {
-			t.Error(err)
-			return
-		}
-
-		parsedIDs = append(parsedIDs, "invalid-tx-id")
-		updatedTxIDStrings, _ := json.Marshal(parsedIDs)
-
-		// discard
-		discardResultsString := DiscardTransactions(C.CString(string(updatedTxIDStrings)))
-		discardResultsStruct := common.DiscardTransactionsResult{}
-		if err := json.Unmarshal([]byte(C.GoString(discardResultsString)), &discardResultsStruct); err != nil {
-			t.Error(err)
-			return
-		}
-		discardResults := discardResultsStruct.Results
-
-		if len(discardResults) != 1 || discardResults["invalid-tx-id"].Error != txqueue.ErrQueuedTxIDNotFound.Error() {
-			t.Errorf("cannot discard txs: %v", discardResults)
-			return
-		}
-
-		// try completing discarded transaction
-		completeResultsString := CompleteTransactions(C.CString(string(updatedTxIDStrings)), C.CString(TestConfig.Account1.Password))
-		completeResultsStruct := common.CompleteTransactionsResult{}
-		if err := json.Unmarshal([]byte(C.GoString(completeResultsString)), &completeResultsStruct); err != nil {
-			t.Error(err)
-			return
-		}
-		completeResults := completeResultsStruct.Results
-
-		if len(completeResults) != (testTxCount + 1) {
-			t.Error("unexpected number of errors (call to CompleteTransaction should not succeed)")
-		}
-		for txID, txResult := range completeResults {
-			if txID != txResult.ID {
-				t.Errorf("tx id not set in result: expected id is %s", txID)
-				return
-			}
-			if txResult.Error != txqueue.ErrQueuedTxIDNotFound.Error() {
-				t.Errorf("invalid error for %s", txResult.Hash)
-				return
-			}
-			if txResult.Hash != zeroHash {
-				t.Errorf("invalid hash (expected zero): %s", txResult.Hash)
-				return
-			}
-		}
-
-		time.Sleep(1 * time.Second) // make sure that tx complete signal propagates
-		for _, txID := range parsedIDs {
-			if txQueue.Has(common.QueuedTxID(txID)) {
-				t.Errorf("txqueue should not have test tx at this point (it should be discarded): %s", txID)
-				return
-			}
-		}
-	}
-	go func() {
-		var txIDStrings []string
-		for i := 0; i < testTxCount; i++ {
-			txIDStrings = append(txIDStrings, <-txIDs)
-		}
-
-		txIDJSON, _ := json.Marshal(txIDStrings)
-		discardTxs(string(txIDJSON))
-	}()
-
-	// send multiple transactions
-	for i := 0; i < testTxCount; i++ {
-		go sendTx()
-	}
-
-	select {
-	case <-allTestTxDiscarded:
-		// pass
-	case <-time.After(20 * time.Second):
-		t.Error("test timed out")
-		return false
-	}
-
-	if txQueue.Count() != 0 {
-		t.Error("tx queue must be empty at this point")
 		return false
 	}
